+++ web/src/screens/Dashboard/Sidebar.tsx
@@ -8,11 +8,15 @@ const Sidebar = () => {
             <div className='mt-10 mx-auto'>
                 <div className='flex flex-col mr-4 pl-0'>
                     {positions.map((position) => (
-                        <div key={position.id} className={`rounded-r my-2 cursor-pointer ${position.id == selectedPosition?.id && 'bg-green-100 border border-l-8 border-l-green-300'}`}
-                        onClick={() => selectPosition(position)}
+                        <div 
+                            key={position.id} 
+                            className={`rounded-r my-2 cursor-pointer ${position.id == selectedPosition?.id && 'bg-green-100 border border-l-8 border-l-green-300'}`}
+                            onClick={() => selectPosition(position)}
                         >
                             <div className='px-4 py-4 sm:px-6'>
-                                <h3 className={`text-lg leading-6 font-medium ${position.id == selectedPosition?.id ? 'text-gray-900' : 'text-gray-600'}`}>{position.name}</h3>
+                                <h3 className={`text-lg leading-6 font-medium ${position.id == selectedPosition?.id ? 'text-gray-900' : 'text-gray-600'}`}>
+                                    {position.name}
+                                </h3>
                             </div>
                         </div>
                     ))}
